Major refactoring.
Create a Node class so that we can re-use it later, and use
collections.ChainMap to handle merging other YAML files more easily.
Also implement a custom PyYAML representer for the Node class, and
output the mapping in a defined order for consistency with Puppet
Dashboard.

The output format is now much closer to Puppet Dashboard's.